refactor management to use shared method after credential resolution
